render table result

+++ src/utils/core/HttpService.js
@@ -103,7 +103,8 @@ class HttpService {
     handleError = error => {
         // Helpers.hideLoading();
         let responstStatus = _get(error, 'response.status')
-        if ( responstStatus === 401 || responstStatus === 400 ) {
+		const pathname = window.location.pathname;
+        if ( responstStatus === 401 || responstStatus === 400 && !pathname.match(/\/auth\/login/i) ) {
             CookieService.removeAll()
             window.history.pushState(null, null, '/auth/login')
             window.location.reload()
